Agents/requirements.txt
accelerate==1.6.0
acres==0.5.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appnope==0.1.4
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.1.1
ci-info==0.3.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cryptography==46.0.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.10.1
debugpy==1.8.13
decorator==5.2.1
deep-translator==1.11.4
Deprecated==1.2.18
dill==0.3.6
distlib==0.3.9
distro==1.9.0
durationpy==0.10
effdet==0.4.1
elastic-transport==8.17.1
elasticsearch==7.13.4
emoji==2.14.1
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
flake8==7.2.0
flatbuffers==25.9.23
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googletrans-py==4.0.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ijson==3.4.0.post0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kubernetes==34.1.0
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.0.10
langchain-huggingface==0.3.1
langchain-tavily==0.2.11
langchain-text-splitters==0.3.9
langdetect==1.0.9
langgraph==0.6.10
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langserve==0.3.1
langsmith==0.4.8
looseversion==1.3.0
lxml==6.0.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.2.0
multiprocess==0.70.14
mypy_extensions==1.1.0
narwhals==2.8.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.19.1
onnxruntime==1.23.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.1
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
patsy==1.0.2
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pi_heif==1.1.1
pikepdf==9.11.0
pillow==11.3.0
platformdirs==4.3.7
pluggy==1.6.0
posthog==5.4.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycodestyle==2.13.0
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.10.1
pydantic_core==2.41.1
pydeck==0.9.1
pydot==4.0.1
pyflakes==3.3.2
Pygments==2.19.1
PyMuPDF==1.26.3
pyogrio==0.10.0
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.8.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproj==3.7.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.3.0
rank-bm25==0.2.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.37.0
regex==2024.11.6
reportlab==4.4.4
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.18.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
sacremoses==0.1.1
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentence-transformers==4.0.2
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.2
statsmodels==0.14.5
streamlit==1.50.0
sympy==1.13.1
tavily-python==0.7.12
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.20
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.50.0
typer==0.19.2
types-python-dateutil==2.9.0.20251008
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
unstructured==0.18.11
unstructured-client==0.41.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.5.0
yarl==1.18.3
zipp==3.23.0
zstandard==0.23.0
